added support for wheel on a non-origin canvas
added babel-cli to the dev-dependencies.
added an optional param to wheel(opts) to allow the plugin to counteract
the offset of the canvas from the html body.

my editor stripped off windows line endings; if the actual code change
looks good enough to merge, I'll make a better pull candidate that
doesn't muck with line endings.